+++ code/system/sys.net.http/src/TEST/libs.mts
@@ -1,0 +1,2 @@
+export * from '../common/index.mjs';
+export { TestFilesystem } from 'sys.fs';
