+++ code/system/sys.net.http/src/TEST/util.mts
@@ -1,8 +1,4 @@
-import { Value } from '../common/index.mjs';
-
-export { expect, expectError } from 'sys.test';
-export { describe, it, beforeEach } from 'vitest';
-export * from '../common/index.mjs';
+import { Value } from './libs.mjs';
 
 /**
  * Generate a random port.
