+++ code/system/sys.net.http/src/common/constants.mts
